shared =============
* enumeration : fix bug introduced in r1394. When historizing an IExtendedEnum, in database there is integers, instead of strings.

The code initially convert the data to string using a boost::lexical_cast (all was working). In r1394, this was changed to use invariant culture, so boost::lexical_cast was replaced by  CStringExtension::cultureInvariantToString (which use std::stringstream to convert data to string). For some reasons, the operator << of std::ostream use the integer implicit cast. This fix now force the behavior of IExtendedEnum used with operator <<